Fixed the conversion function
Fixed the case where someone could input a value using a comma or period to separate thousands and decimals. It isn't perfect but it works if there's both a comma and period in a situation with thousands. just a comma will be interpreted as decimals since it's impossible to discern.